Move plugin actions to endpoint.
